% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.r
\name{babytime}
\alias{babytime}
\title{Calculating time and money based on parental leave and nursery}
\usage{
babytime(
  parent1_wp,
  parent2_wp,
  baby_waking_time,
  baby_sleeping_time,
  parental_leave_option,
  statutory_weekly_pay = 156.66,
  hourly_nursery_rate,
  parent1_annual_salary,
  parent2_annual_salary,
  annual_salary_increment = 0.02
)
}
\arguments{
\item{parent1_wp}{Work plan for parent 1, output from \code{work_plan}}

\item{parent2_wp}{Work plan for parent 2, output from \code{work_plan}}

\item{baby_waking_time}{Expected time that baby wakes up in the morning e.g. "07:00"}

\item{baby_sleeping_time}{Expected time that baby sleeps at night e.g. "19:00"}

\item{parental_leave_option}{Choose option 1 or 2 as per http://www.bristol.ac.uk/hr/policies/shared-parental-leave.html}

\item{statutory_weekly_pay}{Expected weekly pay (as of April 2022 it was 156.66)}

\item{hourly_nursery_rate}{Hourly nursery cost}

\item{parent1_annual_salary}{Gross salary of parent 1 from day 1}

\item{parent2_annual_salary}{Gross salary of parent 2 from day 1}

\item{annual_salary_increment}{Expected annual increment, default=0.02}
}
\value{
List of results
\itemize{
   \item parent1_wp = Parent 1 working times
   \item parent2_wp = Parent 2 working times
   \item baby_schedule = Baby's schedule
   \item overlapping_times = Overlapping times between parents and baby
   \item parent1_salary = Parent 1 salary over time
   \item parent2_salary = Parent 2 salary over time
   \item statutory_parental_pay = Statutory parental pay over time
   \item salary_loss = Salary reductions due to leave
   \item nursery_cost = Nursery costs
   \item weekly_monies = Summary of weekly incomes and outgoings
   \item babytime = Summary of time spent with baby
   \item earnings_plot = Plot of weekly net earnings
   \item babytime_plot = Plot of weekly baby time
}
}
\description{
Calculating time and money based on parental leave and nursery
}
