% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{on_leave}
\alias{on_leave}
\title{Specify a period on shared parental leave}
\usage{
on_leave(weeks)
}
\arguments{
\item{weeks}{Number of weeks for which this period applies}
}
\value{
data frame
}
\description{
Specify a period on shared parental leave
}
