% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{parse_day_time}
\alias{parse_day_time}
\title{Parse day time string}
\usage{
parse_day_time(x)
}
\arguments{
\item{x}{A string such as "Mon-Fri 09:00-17:00" or "Mon,Tue,Wed,Fri 09:00-13:00". Try not to take too many liberties with it, it's not completely comprehensive.}
}
\value{
data frame
}
\description{
Parse day time string
}
