% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baby_time.r
\name{time_with_baby}
\alias{time_with_baby}
\title{Calculate time with baby for one parent}
\usage{
time_with_baby(bwt, wt)
}
\arguments{
\item{bwt}{output from \code{baby_awake_time}}

\item{wt}{output from \code{work_time}}
}
\value{
vector
}
\description{
Some description
}
