% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{work_plan}
\alias{work_plan}
\title{Generate working schedule table}
\usage{
work_plan(start_date, ...)
}
\arguments{
\item{start_date}{First date for table e.g. expected baby's due date}

\item{...}{A list of working or leave periods. e.g. \code{on_leave(26)} signifies 26 weeks of leave, or \code{working(104, "Mon-Thu 09:00-17:00")} signifies 104 weeks of working Monday to Thursday 9-5.}
}
\value{
Data frame
}
\description{
Generate working schedule table
}
