% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.r
\name{working}
\alias{working}
\title{Specify a working period}
\usage{
working(weeks, ...)
}
\arguments{
\item{weeks}{Number of weeks for which this period applies}

\item{...}{The weekly schedule as a string e.g. "Mon-Fri 09:00-17:00" or "Mon,Tue,Wed,Fri 09:00-13:00" that is parsed by \code{parse_day_time}}
}
\value{
data frame
}
\description{
Specify a working period
}
